test(sdk): added tests for token config schemas
